+++ tests/factories.py
@@ -3,7 +3,6 @@ Factory file to create fake datasets of wishlists and wishlists_items for ease i
 """
 
 from datetime import datetime, timezone
-
 from decimal import Decimal
 import factory
 from factory.fuzzy import FuzzyChoice, FuzzyDateTime, FuzzyInteger
