+++ tests/test_wishlist.py
@@ -72,3 +72,104 @@ class TestWishlist(TestCase):
         self.assertEqual(wishlist.created_at, fake_wishlist.created_at)
         self.assertEqual(wishlist.updated_at, fake_wishlist.updated_at)
         self.assertEqual(wishlist.wishlist_items, fake_wishlist.wishlist_items)
+
+
+
+######################################################################
+#        W I S H L I S T  I T E M   M O D E L   T E S T   C A S E S
+######################################################################
+class TestWishlistItem(TestCase):
+    """WishlistItem Model Test Cases"""
+
+    # pylint: disable=duplicate-code
+    @classmethod
+    def setUpClass(cls):
+        """This runs once before the entire test suite"""
+        app.config["TESTING"] = True
+        app.config["DEBUG"] = False
+        app.config["SQLALCHEMY_DATABASE_URI"] = DATABASE_URI
+        app.logger.setLevel(logging.CRITICAL)
+        app.app_context().push()
+
+    # pylint: disable=duplicate-code
+    @classmethod
+    def tearDownClass(cls):
+        """This runs once after the entire test suite"""
+        db.session.close()
+
+    # pylint: disable=duplicate-code
+    def setUp(self):
+        """This runs before each test"""
+        db.session.query(Wishlist).delete()  # clean up the last tests
+        db.session.query(WishlistItem).delete()  # clean up the last tests
+        db.session.commit()
+
+    # pylint: disable=duplicate-code
+    def tearDown(self):
+        """This runs after each test"""
+        db.session.remove()
+
+    ######################################################################
+    #  T E S T   C A S E S
+    ######################################################################
+
+    def test_serialize_an_wishlist_item(self):
+        """It should serialize an wishlist item"""
+        wishlist_item = WishlistItemFactory()
+        serial_wishlist_item = wishlist_item.serialize()
+        self.assertEqual(serial_wishlist_item["id"], wishlist_item.id)
+        self.assertEqual(serial_wishlist_item["wishlist_id"], wishlist_item.wishlist_id)
+        self.assertEqual(serial_wishlist_item["product_id"], wishlist_item.product_id)
+        self.assertEqual(
+            serial_wishlist_item["product_name"], wishlist_item.product_name
+        )
+        self.assertEqual(
+            serial_wishlist_item["product_description"],
+            wishlist_item.product_description,
+        )
+        self.assertEqual(
+            serial_wishlist_item["product_price"], wishlist_item.product_price
+        )
+        self.assertEqual(
+            serial_wishlist_item["created_at"], wishlist_item.created_at.isoformat()
+        )
+        self.assertEqual(
+            serial_wishlist_item["updated_at"],
+            wishlist_item.updated_at.isoformat(),
+        )
+
+    def test_deserialize_a_wishlist_item(self):
+        """It should deserialize a wishlist item"""
+        wishlist_item = WishlistItemFactory()
+        wishlist_item.create()
+        new_wishlist_item = WishlistItem()
+        new_wishlist_item.deserialize(wishlist_item.serialize())
+        self.assertEqual(new_wishlist_item.wishlist_id, wishlist_item.wishlist_id)
+        self.assertEqual(new_wishlist_item.product_id, wishlist_item.product_id)
+        self.assertEqual(new_wishlist_item.product_name, wishlist_item.product_name)
+        self.assertEqual(
+            new_wishlist_item.product_description, wishlist_item.product_description
+        )
+        self.assertEqual(new_wishlist_item.product_price, wishlist_item.product_price)
+
+    def test_add_wishlist_items(self):
+        """It should create a wishlist and assert that it exists"""
+        fake_wishlist = WishlistFactory()
+        # pylint: disable=unexpected-keyword-arg
+        wishlist = Wishlist(
+            name=fake_wishlist.name,
+            id=fake_wishlist.id,
+            is_public=fake_wishlist.is_public,
+            customer_id=fake_wishlist.customer_id,
+            description=fake_wishlist.description,
+            created_at=fake_wishlist.created_at,
+            updated_at=fake_wishlist.updated_at,
+        )
+        self.assertIsNotNone(wishlist)
+        self.assertEqual(wishlist.id, fake_wishlist.id)
+        self.assertEqual(wishlist.name, fake_wishlist.name)
+        self.assertEqual(wishlist.customer_id, fake_wishlist.customer_id)
+        self.assertEqual(wishlist.description, fake_wishlist.description)
+        self.assertEqual(wishlist.created_at, fake_wishlist.created_at)
+        self.assertEqual(wishlist.updated_at, fake_wishlist.updated_at)
+        self.assertEqual(wishlist.wishlist_items, fake_wishlist.wishlist_items)
